requirements.txt
opencv-python==4.0.0.21
numpy==1.16.2
aido-protocols-daffy>=5.0.16
duckietown-world-daffy>=5.0.7
duckietown-challenges-daffy>=5.0.7
pillow
torch
torchvision
